+++ src/BubbleScript.js
@@ -6,9 +6,6 @@
      *        *    +   *
        *  */
 
-
-let emptyList, emptyVector;
-
 const keywords    = {};
 const symbols     = {};
 
@@ -21,47 +18,7 @@ const TOK_KEYWORD = 'K',
 
 const AbstractList = require("./o/abstract_list");
 const List = require("./o/list");
-
-class Vector extends AbstractList {
-
-  static get emptyVector() { return emptyVector; }
-
-  static make(...elements) {
-    var head = emptyVector;
-    for (let o of elements)
-      head = new this(o, head);
-    return head;
-  }
-
-  constructor(o, oo=emptyVector) {
-    super(o, oo);
-  }
-
-  push(element) {
-    return new Vector(element, this);
-  }
-
-  toString() {
-    return "[" + this._toString() + "]";
-  }
-
-  toStringJoin(accumulatedString, formattedElement) {
-    return formattedElement + " " + accumulatedString;
-  };
-
-  toList() {
-    return this.reduce((list, o) => {
-      return list.push(o); },
-      List.emptyList);
-  }
-
-}
-
-class EmptyVector extends Vector {
-  get isEmpty() { return true; }
-}
-
-emptyVector = new EmptyVector()
+const Vector = require("./o/vector");
 
 
 // `§ymbol`s are language symbols. Underscored
